ITD/backend/requirements.txt
flask_restful
Flask-Cors
bcrypt
pyjwt
python-dotenv